src/ledger-telegram-bot/src/requirements.txt
python-telegram-bot==13.15
dateparser==1.1.8
